2030-04-02 / 1:13AM

+++ src/main/java/lk/ijse/sports_zone/controller/AdminEmployeeFormController.java
@@ -122,22 +122,24 @@ public class AdminEmployeeFormController {
 
     @FXML
     void deleteOnAction(ActionEvent event) {
-        btnSave.setVisible(true);
-        try {
-            String empId = txtEmpId.getText();
+        boolean result = AlertController.okconfirmmessage("Are you sure you want to delete ?");
+        if(result) {
+            btnSave.setVisible(true);
+            try {
+                String empId = txtEmpId.getText();
 
-            boolean isDeleted = EmployeeModel.delete(empId);
-            if(isDeleted){
-                AlertController.successfulMessage("Deleted");
-                setCellValueFactory();
-                getAll();
-                clearTxtField();
-            }else {
-                AlertController.errormessage("Not Deleted");
+                boolean isDeleted = EmployeeModel.delete(empId);
+                if (isDeleted) {
+                    setCellValueFactory();
+                    getAll();
+                    clearTxtField();
+                } else {
+                    AlertController.errormessage("Not Deleted");
+                }
+            } catch (SQLException throwables) {
+                throwables.printStackTrace();
+                AlertController.exceptionMessage("Something went wrong");
             }
-        } catch (SQLException throwables) {
-            throwables.printStackTrace();
-            AlertController.exceptionMessage("Something went wrong");
         }
     }
 
@@ -166,17 +168,14 @@ public class AdminEmployeeFormController {
 
                         boolean isSaved = EmployeeModel.save(employee);
                         if(isSaved){
-                            AlertController.successfulMessage("Saved");
-
                             setCellValueFactory();
                             getAll();
                             clearTxtField();
+                            AlertController.successfulMessage("Saved");
                         }
-                    } catch (SQLException throwables) {
-                        throwables.printStackTrace();
-                        AlertController.exceptionMessage("SQLException");
-                    }catch(Exception exception){
+                    } catch(Exception exception){
                         AlertController.exceptionMessage("Something went wrong");
+                        System.out.println("EmpSave ="+exception);
                     }
 
                 } else {
@@ -194,52 +193,56 @@ public class AdminEmployeeFormController {
     @FXML
     void updateOnAction(ActionEvent event) {
         //Employee employee = new Employee();
-        btnSave.setVisible(true);
+        boolean result = AlertController.okconfirmmessage("Are you sure you want to delete ?");
+        if(result){
+            btnSave.setVisible(true);
 
-        if (ValidateController.emailCheck(txtEmail.getText()) || ValidateController.contactCheck(txtContactNo.getText())) {
-            if (ValidateController.contactCheck(txtContactNo.getText())) {
-                //lblInvalidContacktNo.setVisible(false);
+            if (ValidateController.emailCheck(txtEmail.getText()) || ValidateController.contactCheck(txtContactNo.getText())) {
+                if (ValidateController.contactCheck(txtContactNo.getText())) {
+                    //lblInvalidContacktNo.setVisible(false);
 
-                if (ValidateController.emailCheck(txtEmail.getText())) {
-                    //lblInvalidEmail.setVisible(false);
+                    if (ValidateController.emailCheck(txtEmail.getText())) {
+                        //lblInvalidEmail.setVisible(false);
 
-                    employee.setEmpId(txtEmpId.getText());
-                    employee.setEmpName(txtEmpName.getText());
-                    employee.setAddress(txtAddress.getText());
-                    employee.setDob(String.valueOf(txtDob.getValue()));
-                    employee.setContactNo(txtContactNo.getText());
-                    employee.setSalary(Double.valueOf(txtSalary.getText()));
-                    employee.setEmail(txtEmail.getText());
-                    employee.setNic(txtNIC.getText());
-                    employee.setJobTitle(cmbJobTitle.getValue());
+                        employee.setEmpId(txtEmpId.getText());
+                        employee.setEmpName(txtEmpName.getText());
+                        employee.setAddress(txtAddress.getText());
+                        employee.setDob(String.valueOf(txtDob.getValue()));
+                        employee.setContactNo(txtContactNo.getText());
+                        employee.setSalary(Double.valueOf(txtSalary.getText()));
+                        employee.setEmail(txtEmail.getText());
+                        employee.setNic(txtNIC.getText());
+                        employee.setJobTitle(cmbJobTitle.getValue());
 
-                    try {
-                        boolean isUpdated = EmployeeModel.update(employee);
-                        if(isUpdated){
-                            AlertController.successfulMessage("updated");
-                            setCellValueFactory();
-                            getAll();
-                            clearTxtField();
-                            btnSave.setDisable(false);
-                        }else {
-                            AlertController.errormessage("Not Updated");
+                        try {
+                            boolean isUpdated = EmployeeModel.update(employee);
+                            if(isUpdated){
+                                AlertController.okMassage("updated");
+                                setCellValueFactory();
+                                getAll();
+                                clearTxtField();
+                                btnSave.setDisable(false);
+                            }else {
+                                AlertController.errormessage("Not Updated");
+                            }
+
+                        } catch (SQLException throwables) {
+                            throwables.printStackTrace();
+                            AlertController.exceptionMessage("Something went wrong");
                         }
 
-                    } catch (SQLException throwables) {
-                        throwables.printStackTrace();
-                        AlertController.exceptionMessage("Something went wrong");
+                    } else {
+                        lblInvalidEmail.setVisible(true);
                     }
-
                 } else {
-                    lblInvalidEmail.setVisible(true);
+                    lblInvalidContacktNo.setVisible(true);
                 }
             } else {
+                lblInvalidEmail.setVisible(true);
                 lblInvalidContacktNo.setVisible(true);
             }
-        } else {
-            lblInvalidEmail.setVisible(true);
-            lblInvalidContacktNo.setVisible(true);
         }
+
     }
